+++ sdk/communication/AzureCommunicationChat/Source/Signaling/Events/ChatEvent.swift
@@ -78,8 +78,6 @@ public class BaseChatEvent {
     public var threadId: String
     /// Sender identifier.
     public var sender: CommunicationIdentifier?
-    /// Sender display name.
-    public var senderDisplayName: String?
     /// Recipient identifier.
     public var recipient: CommunicationIdentifier?
     
@@ -89,17 +87,14 @@ public class BaseChatEvent {
     /// - Parameters:
     ///   - threadId: ChatThread id.
     ///   - sender: Sender identifier.
-    ///   - senderDisplayName: Sender display name.
     ///   - recipient: Recipient identifier.
     init(
         threadId: String,
         sender: CommunicationIdentifier?,
-        senderDisplayName: String? = nil,
         recipient: CommunicationIdentifier?
     ) {
         self.threadId = threadId
         self.sender = sender
-        self.senderDisplayName = senderDisplayName
         self.recipient = recipient
     }
 }
@@ -137,6 +132,8 @@ public class BaseChatMessageEvent: BaseChatEvent {
     public var version: String
     /// The message type.
     public var type: ChatMessageType
+    /// Sender display name.
+    public var senderDisplayName: String?
     
     // MARK: Initializers
     
@@ -164,7 +161,8 @@ public class BaseChatMessageEvent: BaseChatEvent {
         self.createdOn = createdOn
         self.version = version
         self.type = type
-        super.init(threadId: threadId, sender: sender, senderDisplayName: senderDisplayName, recipient: recipient)
+        self.senderDisplayName = senderDisplayName
+        super.init(threadId: threadId, sender: sender, recipient: recipient)
     }
 }
 
@@ -299,10 +297,12 @@ public class ChatMessageEditedEvent: BaseChatMessageEvent {
 
         self.message = chatMessageEditedPayload.messageBody
         self.editedOn = Iso8601Date(string: chatMessageEditedPayload.edittime)
+        // TODO: Hardcode "8:" as workaround to missing prefix in payload
+        let recipientId = "8:\(chatMessageEditedPayload.recipientId)"
         super.init(
             threadId: chatMessageEditedPayload.groupId,
             sender: TrouterEventUtil.getIdentifier(from: chatMessageEditedPayload.senderId),
-            recipient: TrouterEventUtil.getIdentifier(from: chatMessageEditedPayload.recipientId),
+            recipient: TrouterEventUtil.getIdentifier(from: recipientId),
             id: chatMessageEditedPayload.messageId,
             senderDisplayName: chatMessageEditedPayload.senderDisplayName,
             createdOn: Iso8601Date(string: chatMessageEditedPayload.originalArrivalTime),
@@ -367,10 +367,12 @@ public class ChatMessageDeletedEvent: BaseChatMessageEvent {
             .decode(MessageDeletedPayload.self, from: requestJsonData)
 
         self.deletedOn = Iso8601Date(string: chatMessageDeletedPayload.deletetime)
+        // TODO: Hardcode "8:" as workaround to missing prefix in payload
+        let recipientId = "8:\(chatMessageDeletedPayload.recipientId)"
         super.init(
             threadId: chatMessageDeletedPayload.groupId,
             sender: TrouterEventUtil.getIdentifier(from: chatMessageDeletedPayload.senderId),
-            recipient: TrouterEventUtil.getIdentifier(from: chatMessageDeletedPayload.recipientId),
+            recipient: TrouterEventUtil.getIdentifier(from: recipientId),
             id: chatMessageDeletedPayload.messageId,
             senderDisplayName: chatMessageDeletedPayload.senderDisplayName,
             createdOn: Iso8601Date(string: chatMessageDeletedPayload.originalArrivalTime),
@@ -423,11 +425,12 @@ public class TypingIndicatorReceivedEvent: BaseChatEvent {
 
         self.version = typingIndicatorReceivedPayload.version
         self.receivedOn = Iso8601Date(string: typingIndicatorReceivedPayload.originalArrivalTime)
-
+        // TODO: Hardcode "8:" as workaround to missing prefix in payload
+        let recipientId = "8:\(typingIndicatorReceivedPayload.recipientId)"
         super.init(
             threadId: typingIndicatorReceivedPayload.groupId,
             sender: TrouterEventUtil.getIdentifier(from: typingIndicatorReceivedPayload.senderId),
-            recipient: TrouterEventUtil.getIdentifier(from: typingIndicatorReceivedPayload.recipientId)
+            recipient: TrouterEventUtil.getIdentifier(from: recipientId)
         )
     }
 }
@@ -479,15 +482,24 @@ public class ReadReceiptReceivedEvent: BaseChatEvent {
         let readReceiptMessageBody: ReadReceiptMessageBody = try JSONDecoder()
             .decode(ReadReceiptMessageBody.self, from: readReceiptMessageBodyJsonData)
 
+        // Extract readOn value from consumptionHorizon
         let consumptionHorizon = readReceiptMessageBody.consumptionhorizon.split(separator: ";")
-        let readOn = String(consumptionHorizon[1])
+        guard let readOnMs = Double(consumptionHorizon[1]) else {
+            throw AzureError.client("Failed to construct Int from consumptionHorizon for readOn property.")
+        }
+
+        // In the payload readOn is represented as epoch time in milliseconds 
+        let readOnSeconds = readOnMs / 1000
+        let readOnDate = Date(timeIntervalSince1970: TimeInterval(readOnSeconds))
 
         self.chatMessageId = readReceiptReceivedPayload.messageId
-        self.readOn = Iso8601Date(string: readOn)
+        self.readOn = Iso8601Date(readOnDate)
+        // TODO: Hardcode "8:" as workaround to missing prefix in payload
+        let recipientId = "8:\(readReceiptReceivedPayload.recipientId)"
         super.init(
             threadId: readReceiptReceivedPayload.groupId,
             sender: TrouterEventUtil.getIdentifier(from: readReceiptReceivedPayload.senderId),
-            recipient: TrouterEventUtil.getIdentifier(from: readReceiptReceivedPayload.recipientId)
+            recipient: TrouterEventUtil.getIdentifier(from: recipientId)
         )
     }
 }